refactor: 이력서 -> 자기소개서
