+++ src/main/java/com/aibe/team2/domain/statistics/enums/ServiceType.java
@@ -6,7 +6,7 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public enum ServiceType {
-    RESUME("이력서 첨삭"),
+    RESUME("자기소개서 첨삭"),
     INTERVIEW("모의 면접");
 
     private final String description;
